[Mod] WORK_TYPE.java enum class separate code from MetaExtractWorker.java
