+++ db2pg/src/main/java/com/k4m/experdb/db2pg/rebuild/TargetPgDDL.java
@@ -11,7 +11,7 @@ import com.k4m.experdb.db2pg.db.DBCPPoolManager;
 import com.k4m.experdb.db2pg.db.datastructure.DBConfigInfo;
 import com.k4m.experdb.db2pg.work.db.impl.MetaExtractWork;
 import com.k4m.experdb.db2pg.work.db.impl.MetaExtractWorker;
-import com.k4m.experdb.db2pg.work.db.impl.MetaExtractWorker.WORK_TYPE;
+import com.k4m.experdb.db2pg.work.db.impl.WORK_TYPE;
 
 public class TargetPgDDL {
 	private List<String> idxCreateList;
