+++ db2pg/src/main/java/com/k4m/experdb/db2pg/work/db/impl/MetaExtractWork.java
@@ -2,8 +2,6 @@ package com.k4m.experdb.db2pg.work.db.impl;
 
 import java.util.Map;
 
-import com.k4m.experdb.db2pg.work.db.impl.MetaExtractWorker.WORK_TYPE;
-
 public class MetaExtractWork {
 	WORK_TYPE type;
 	Map<String,Object> params;
