+++ db2pg/src/main/java/com/k4m/experdb/db2pg/work/db/impl/MetaExtractWorker.java
@@ -106,8 +106,4 @@ public final class MetaExtractWorker extends DBWorker {
 		return poolName;
 	}
 
-	public enum WORK_TYPE {
-		GET_TABLE_NAMES, GET_TABLE_INFORM, GET_COLUMN_INFORM, GET_CONSTRAINT_INFORM, GET_KEY_INFORM, GET_AUTOINCREMENT_INFORM, GET_PG_CURRENT_SCHEMA, GET_PG_IDX_DDL, GET_PG_FK_DDL, GET_VIEW_INFORM, GET_SEQUENCE_INFORM
-	}
-
 }
